+++ Sources/MapboxMaps/Location/Pucks/Puck3D.swift
@@ -78,7 +78,7 @@ internal class Puck3D: Puck {
 
             // TODO: On first setup "puck-model does not have a uri"
             try? style.addSource(self.modelSource, id: "puck-model-source")
-            try! style._addPersistentLayer(self.modelLayer)
+            try! style.addPersistentLayer(self.modelLayer)
         }
 
         // Do initial setup