+++ Sources/MapboxMaps/Style/StyleManagerProtocol.swift
@@ -74,15 +74,17 @@ internal protocol StyleManagerProtocol {
     /// - Throws:
     ///     An error describing why the operation was unsuccessful
     ///
-    /// - Note: This API is experimental and can change at any time
-    func _addPersistentLayer(with properties: [String: Any], layerPosition: LayerPosition?) throws
+    /// - Note: This API is experimental and can change at any time. Annotate the import statement
+    /// for `MapboxMaps` with `@_spi(Experimental)` in order to use experimental methods.
+    func addPersistentLayer(with properties: [String: Any], layerPosition: LayerPosition?) throws
 
     /// Returns `true` if the id passed in is associated to a persistent layer
     ///
-    /// - Note: This API is experimental and can change at any time
+    /// - Note: This API is experimental and can change at any time. Annotate the import statement
+    /// for `MapboxMaps` with `@_spi(Experimental)` in order to use experimental methods.
     ///
     /// - Parameter id: The layer identifier to test
-    func _isPersistentLayer(id: String) throws -> Bool
+    func isPersistentLayer(id: String) throws -> Bool
 
     /// Adds a new persistent style custom layer.
     ///
@@ -99,8 +101,9 @@ internal protocol StyleManagerProtocol {
     /// - Throws:
     ///     An error describing why the operation was unsuccessful.
     ///
-    /// - Note: This API is experimental and can change at any time
-    func _addPersistentCustomLayer(withId id: String, layerHost: CustomLayerHost, layerPosition: LayerPosition?) throws
+    /// - Note: This API is experimental and can change at any time. Annotate the import statement
+    /// for `MapboxMaps` with `@_spi(Experimental)` in order to use experimental methods.
+    func addPersistentCustomLayer(withId id: String, layerHost: CustomLayerHost, layerPosition: LayerPosition?) throws
 
     /// Adds a new style custom layer.
     ///
@@ -153,7 +156,9 @@ internal protocol StyleManagerProtocol {
     ///
     /// - Returns:
     ///     The value of the property in the layer with layerId.
-    func _layerProperty(for layerId: String, property: String) -> StylePropertyValue
+    /// - Note: This method is marked as experimental. Annotate the import statement
+    /// for `MapboxMaps` with `@_spi(Experimental)` in order to use experimental methods.
+    func layerProperty(for layerId: String, property: String) -> StylePropertyValue
 
     /// Sets a JSON value to a style layer property.
     ///
@@ -176,7 +181,10 @@ internal protocol StyleManagerProtocol {
     ///
     /// - Returns:
     ///     The default value of the property for the layers with type layerType.
-    static func _layerPropertyDefaultValue(for layerType: LayerType, property: String) -> StylePropertyValue
+    ///
+    /// - Note: This method is marked as experimental. Annotate the import statement
+    /// for `MapboxMaps` with `@_spi(Experimental)` in order to use experimental methods.
+    static func layerPropertyDefaultValue(for layerType: LayerType, property: String) -> StylePropertyValue
 
     /// Gets the properties for a style layer.
     ///
@@ -250,7 +258,10 @@ internal protocol StyleManagerProtocol {
     ///   - property: Style source property name.
     ///
     /// - Returns: The value of the property in the source with sourceId.
-    func _sourceProperty(for sourceId: String, property: String) -> StylePropertyValue
+    ///
+    /// - Note: This method is marked as experimental. Annotate the import statement
+    /// for `MapboxMaps` with `@_spi(Experimental)` in order to use experimental methods.
+    func sourceProperty(for sourceId: String, property: String) -> StylePropertyValue
 
     /// Sets a value to a style source property.
     ///
@@ -300,7 +311,10 @@ internal protocol StyleManagerProtocol {
     ///
     /// - Returns:
     ///     The default value for the named property for the sources with type sourceType.
-    static func _sourcePropertyDefaultValue(for sourceType: String, property: String) -> StylePropertyValue
+    ///
+    /// - Note: This method is marked as experimental. Annotate the import statement
+    /// for `MapboxMaps` with `@_spi(Experimental)` in order to use experimental methods.
+    static func sourcePropertyDefaultValue(for sourceType: String, property: String) -> StylePropertyValue
 
     // MARK: Image source
 
@@ -386,7 +400,10 @@ internal protocol StyleManagerProtocol {
     /// - Parameter property: Style light property name.
     ///
     /// - Returns: Style light property value.
-    func _lightProperty(_ property: String) -> StylePropertyValue
+    ///
+    /// - Note: This method is marked as experimental. Annotate the import statement
+    /// for `MapboxMaps` with `@_spi(Experimental)` in order to use experimental methods.
+    func lightProperty(_ property: String) -> StylePropertyValue
 
     /// Sets a value to the style light property.
     ///
@@ -416,7 +433,10 @@ internal protocol StyleManagerProtocol {
     /// - Parameter property: Style terrain property name.
     ///
     /// - Returns: Style terrain property value.
-    func _terrainProperty(_ property: String) -> StylePropertyValue
+    ///
+    /// - Note: This method is marked as experimental. Annotate the import statement
+    /// for `MapboxMaps` with `@_spi(Experimental)` in order to use experimental methods.
+    func terrainProperty(_ property: String) -> StylePropertyValue
 
     /// Sets a value to the named style terrain property.
     ///
@@ -453,8 +473,11 @@ internal protocol StyleManagerProtocol {
     ///   - features: An array of features to add
     ///
     /// - Throws:
-    ///     An error describing why the operation was unsuccessful.\
-    func _setCustomGeometrySourceTileData(forSourceId sourceId: String, tileId: CanonicalTileID, features: [Turf.Feature]) throws
+    ///     An error describing why the operation was unsuccessful.
+    ///
+    /// - Note: This method is marked as experimental. Annotate the import statement
+    /// for `MapboxMaps` with `@_spi(Experimental)` in order to use experimental methods.
+    func setCustomGeometrySourceTileData(forSourceId sourceId: String, tileId: CanonicalTileID, features: [Turf.Feature]) throws
 
     /// Invalidate tile for provided custom geometry source.
     ///